requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
psycopg2-binary==2.9.9
voyageai==0.3.2
anthropic==0.34.0
pydantic==2.8.2
